[ADD] Test JS OK & Test REACT OK
