--- react/test2_react/src/App.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import './App.css';
-import Counter from './component/Counter';
-
-function App() {
-  return (
-    <div className="App">
-      <h1>Incrémenter ou Décrémenter :) !</h1>
-      <Counter />
-    </div>
-  );
-}
-
-export default App;
