+++ react/test_react/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "test2_react",
+  "name": "test_react",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
