+++ react/test_react/src/component/Counter.js
@@ -1,13 +1,12 @@
 import React, { useState } from 'react';
 
 function Counter() {
-
   const [count, setCount] = useState(0);
 
   const increment = () => {
     setCount(count + 1);
   };
-  
+
   const decrement = () => {
     setCount(count - 1);
   };
